+++ tests/functions/Search-TssSecret.Tests.ps1
@@ -4,8 +4,15 @@
 }
 Describe "$commandName veirfy parameters" {
     BeforeDiscovery {
-        [object[]]$knownParameters = 'TssSession', 'FolderId', 'SecretTemplateId', 'SiteId', 'HeartbeatStatus', 'SearchField', 'SearchText', 'SearchSlug', 'IncludeInactive', 'Raw'
+        [object[]]$knownParameters = 'TssSession','Raw','Permission','Scope',
+        # Folder param set
+        'FolderId','IncludeSubFolders',
+        # field param set
+        'Field','FieldText','ExactMatch','FieldSlug','ExtendedField','ExtendedTypeId',
+        # secret param set
+        'SecretTemplateId','SiteId','HeartbeatStatus','IncludeInactive','ExcludeActive','RpcEnabled','SharedWithMe','PasswordTypeIds','ExcludeDoubleLock','DoubleLockId'
         [object[]]$currentParams = ([Management.Automation.CommandMetaData]$ExecutionContext.SessionState.InvokeCommand.GetCommand($commandName, 'Function')).Parameters.Keys
+        [object[]]$commandDetails = [System.Management.Automation.CommandInfo]$ExecutionContext.SessionState.InvokeCommand.GetCommand($commandName,'Function')
         $unknownParameters = Compare-Object -ReferenceObject $knownParameters -DifferenceObject $currentParams -PassThru
     }
     Context "Verify parameters" -Foreach @{currentParams = $currentParams} {
@@ -14,6 +21,51 @@ Describe "$commandName veirfy parameters" {
         }
         It "$commandName should not contain parameter: <_>" -TestCases $unknownParameters {
             $_ | Should -BeNullOrEmpty
+        }
+    }
+    Context "Command specific details" {
+        It "$commandName should set OutputType to TssSecretSummary" -TestCases $commandDetails {
+            $_.OutputType.Name | Should -Be 'TssSecretSummary'
+        }
+    }
+}
+Describe "$commandName works" {
+    BeforeDiscovery {
+        $session = New-TssSession -SecretServer $ss -Credential $ssCred
+        $object = Search-TssSecret $session -RpcEnabled -SecretTemplateId 6003
+        $session.SessionExpire()
+
+        $props = 'SecretId','FolderId','SecretTemplateId','Name'
+    }
+    Context "Checking" -Foreach @{object = $object} {
+        It "Should not be empty" {
+            $object | Should -Not -BeNullOrEmpty
+        }
+        It "Should find one secret with RPC enabled" {
+            $object.SecretTemplateId | Should -Be 6003
+        }
+        It "Should output <_> property" -TestCases $props {
+            $object.PSObject.Properties.Name | Should -Contain $_
+        }
+    }
+}
+Describe "$commandName works" {
+    BeforeDiscovery {
+        $session = New-TssSession -SecretServer $ss -Credential $ssCred
+        $object = Search-TssSecret $session -RpcEnabled -SecretTemplateId 6003
+        $session.SessionExpire()
+
+        $props = 'SecretId','FolderId','SecretTemplateId','SecretName'
+    }
+    Context "Checking" -Foreach @{object = $object} {
+        It "Should not be empty" {
+            $object | Should -Not -BeNullOrEmpty
+        }
+        It "Should find one secret with RPC enabled" {
+            $object.SecretTemplateId | Should -Be 6003
+        }
+        It "Should output <_> property" -TestCases $props {
+            $object.PSObject.Properties.Name | Should -Contain $_
         }
     }
 }